Improved text
I replaced the colon in "the beautiful coastal city: Manzanillo" for "of". 
Removed capitals in "Attendees", "Conferences", and "Keynotes".
Removed "Conferences" in "Conferences speakers".
Replaced "series of houses" for "handful of houses". 
Added a comma after "houses". 
Added "all" before "put together". 
Deleted the "s" in "conferences in "Development conferences"
Added a period at the end of Sandy Metz's bio.